updating to remove comments, clean up per SG comments
